Separate code for diagnose
